complete export_events refactor & tests
